refactor: refactor Comment and rename functions
