+++ apps/web/src/features/comments/actions/deleteCommentAction.ts
@@ -3,7 +3,7 @@
 import { trpc } from '@/fsd/shared/trpc/lib/server';
 import { revalidatePath } from 'next/cache';
 
-export async function deleteComment(commentId: string, postId: string) {
+export async function deleteCommentAction(commentId: string, postId: string) {
 	await trpc.comment.deleteComment(commentId);
 
 	revalidatePath(`/blog/${postId}`);
