+++ apps/web/src/features/comments/actions/updateCommentAction.ts
@@ -3,7 +3,7 @@
 import { trpc } from '@/fsd/shared/trpc/lib/server';
 import { revalidatePath } from 'next/cache';
 
-export async function updateComment(
+export async function updateCommentAction(
 	commentId: string,
 	content: string,
 	postId: string,
